[Compilation] Fix cmake build for < GCC9, Linux
On linux, with GCC 8.4.1 and cmake 3.18.2, I observe:

```
...
[ 94%] Linking C static library ../../lib/libSpdmDeviceSecretLib.a
/bin/ranlib: ../../lib/libSpdmDeviceSecretLib.a(SpdmDeviceSecretLib.c.o): plugin needed to handle lto object
/bin/ranlib: ../../lib/libSpdmDeviceSecretLib.a(SpdmDeviceSecretLibCert.c.o): plugin needed to handle lto object

...
/tmp/ccifeBRc.ltrans0.ltrans.o: In function `SpdmClientInit':
<artificial>:(.text+0x29e): undefined reference to `SpdmGetContextSize'
<artificial>:(.text+0x301): undefined reference to `SpdmInitContext'
<artificial>:(.text+0x32d): undefined reference to `SpdmRegisterDeviceIoFunc'
```

Some cursory searching suggests the fix is
https://stackoverflow.com/q/39236917.

As the post explains, the Cmake template for c libraries forces an
invocation of `ranlib`, which adds an index to archives. However, the
hard coded ranlib invocation inside cmake source does not invoke the
plugin, which causes ranlib to fail (somewhat ungracefully, it seems).

However, the compilation will continue, and later steps in linking will
try to use the "index" generated by ranlib, and predictably fail, since
ranlib never ran properly.

My fix for this is a simplified version of what the StackOverflow
recommends. We can simply omit ranlib when it is not known to work
correctly by overriding CMAKE_RANLIB with a noop. For now, I've pinned
it to 9.3.0, since we know that works based on jyao1's testing. In the
future, if we can find tighter bounds it can be updated.

The reason to add this branch is because ranlib is helpful to the link
speed of the binary. This isn't a huge deal since openspdm is fairly
small, but no reason to introduce slowdown unless impeding correctness.

Testing reveals minor impact to link time (<0.5s), basically noise.

Signed-off-by: Howard McLauchlan <hmclauchlan@fb.com>
Signed-off-by: Jiewen Yao <jiewen.yao@intel.com>