record transcript hash instead of data.
REF: https://github.com/DMTF/libspdm/issues/22

The context size is reduced from 0x12c10 (76K) to 0xf3d8 (62K).

The scope is limited to message_b/c/mut_b/mut_c and
message_m(measurement).

We cannot do for message_a, because the algo is not negotiated yet.

We have not done for message_k/f(key_exchange/finish) yet.
That will be done in next patch set.

Test:
1) Pass unit test with RECORD_TRANSCRIPT_DATA on/off
2) Pass emu test with responder (data) + requester (hash)
3) Pass emu test with responder (hash) + requester (data)

Signed-off-by: Jiewen Yao <jiewen.yao@intel.com>