Add new line at end of files
Some compiler warning configurations will emit a warning if a new line
is not detected at the end of a file.

This change makes all the files consistent and adds a new line to the
end of every file.

A follow on step is needed to protect this by including a compliance
check in CI.

Signed-off-by: Abe Kohandel <abe.kohandel@intel.com>