Initial pass at cleaning up crypto library builds
This change removes the unused cryptstublib_dummy as a step to start
simplifying the crypto library builds to decouple them from the main
library builds.

There is only ever a single crypto library used, there is no need for
naming them differently. This simplifies the build process by just
depending on a cryptlib library that is either provided by an openssl
implementation or an embedtls implementation.

There is a cryptlib_null still that is used in the size tests but that
can be ignored as it is not really a functional crypto library.

mbedtls can also be used to build without compiling from source but the
tests fail.

Signed-off-by: Abe Kohandel <abe.kohandel@intel.com>