We added unit test cases for the key_exchange messages, expanding the existing tests for the message.
These tests verify the device behavior when receiving measurement hashes, heartbeat periods, mutual authentication requests and different certificate slots requests. There is also a test in which a wrong signature is provided to check if the device deals correctly with failures in signature validation.

For these tests, we included a variable to be used as a TCB hash value in spdm_unit_test_common/algo.c.

Additionally, we corrected a minor error in libspdm_rsp_key_exchange.c: on line 128, the code was checking the requester param2 variable when it should check param1.

In this push request:

1) Additional unit test cases for key_exchange requester (17 new tests)

- cases 12 to 17 are related to measurement hashes

- case 18 is related to the wrong signature validation

- case 19 is related to the ResponderVerifyData field

- cases 20 to 22 are related to the heartbeat period

- cases 23 to 28 are related to mutual authentication scenarios

2) Additional unit test cases for key_exchange responder (8 new tests)

- cases 8 to 11 are related to measurement hashes

- cases 12 to 14 are related to certificates and slot ids

- case 15 is related to the HANDSHAKE_IN_THE_CLEAR flag

3) Modification to the key_exchange (responder) library:

- (line 128) Checks if param1 is set as SPDM_CHALLENGE_REQUEST_NO_MEASUREMENT_SUMMARY_HASH

Signed-off-by: Eduardo Lopes Cominetti <ecominetti@larc.usp.br>