libspdm_rsp_psk_exchange: Fixup PSK Hint Length comparison
The maximum length of the PSK hint is referenced by the
LIBSPDM_PSK_MAX_HINT_LENGTH macro (16). We need to ensure that the
length is less then equal to the length of LIBSPDM_PSK_MAX_HINT_LENGTH,
not less then.

This patch ensures that the psk hint is less then or equal to
LIBSPDM_PSK_MAX_HINT_LENGTH which also matches the check when setting
the psk_hint buffer for the local context.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>