Supply slot_id to libspdm_*_data_sign()
Supply the slot_id to libspdm_responder_data_sign() and
libspdm_requester_data_sign(). These functions use certificates to sign
an SPDM message so they need to know what slot is being used to sign the
data.

For CHALLENGE_AUTH for example the slot_id is sent as part of Param1 and
section 365 of spec version 1.2.1 specifies that that slot leaf
certificate should be used for signing. Currently there is no way to
know which leaf certificate should be used when generating signatures,
this patch ensures that we know what slot certificate should be used.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>