library: Include base_*_algo to libspdm_write_certificate_to_nvm()
If the implementor wants to update the certificate without rebooting the
device they will need to know the base_hash_algo and base_asym_algo in
order to verify and hash the new certificate. Let's pass that
information to the libspdm_write_certificate_to_nvm() function so that
it can use it.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>