lib: libspdm_rsp_capabilities: fixup spec conformance
When chunking is not supported, MaxSPDMmsgSize shall equal
DataTransferSize in a GET_CAPABILITIES request. However,
this is only specified in the SPDM 1.3 spec.

A requester implementing 1.2 >= version < 1.3 may not conform to this as
it is not specified in the 1.2 or 1.2.1 specification for a
GET_CAPABILITIES request. So only enfore this if the version used is
1.3.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>