cryptlib_openssl: x509: fixup retval
X509_verify_cert() returns 1 if a complete chain can be built and
validated, otherwise 0, and in exceptional circumstances
(such as malloc failure and internal errors) they can also return a
negative code [1].

By casting the int return value to a bool, a negative error code
can be treated as `true`. This change fixes that.

[1] https://www.openssl.org/docs/man3.0/man3/X509_verify_cert.html

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>