crypt_lib: allow partial certchain verification
The new addition to the changed api allows us to verify a partial
certificate chain, that is, one that does not contain a leaf
certificate.

This maybe useful, for example, when the requester is doing a
`SET_CERTIFICATE` and is loading in the immutable certificates. If using
the AliasCert model, then this chain contains only the rootCA->...->deviceCA
(the `immutable` certificates). Composed of only CAs (no leaf). In such a
case, we can still use this function to validate most of the chain,
but skip the leaf check.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>