crypt_lib: implement new api with partial verify
This adds a new function `libspdm_verify_cert_chain_data_ex()`. It
differs from `libspdm_verify_cert_chain_data()` in that, it should be
used to verify a certchain for data exchange, e.x prior to SET_CERTIFICATE.
This function takes an extra argument `check_leaf_cert` that allows
for skipping the verification of the leaf certificate in the
certificate chain.

This maybe useful for a libspdm requester to validate the local
immutable certificate chain prior to a `SET_CERTIFICATE` request. That
is, when using the AliasCertificate model, the chain will not contain a leaf
certificate. So we can skip that check, but still validate the rest of
the chain.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>