os_stub: cryptlib: CSR: Allow copying attributes from an existing cert
Section 754 of SPDM 1.3 states that:

    The CSR shall be populated with a combination of attributes
    provided by the Requester via the RequesterInfo field and other
    attributes contributed by the Responder itself.

libspdm currently supports populating the CSR based on the
RequesterInfo, but we don't allow the Responder to add custom
attributes.

This patch updated libspdm_gen_x509_csr() to pass in a base certificate
that is used to provide CSR attributes. libspdm_gen_x509_csr() will copy
the extensions from base_cert to the CSR.

This way devices can choose to apply the same extensions from a leaf
certificate (or other certificate already on the device) or create a new
cert just to pass the extensions in.

By copying the extensions from an existing cert the API should be
generic between OpenSSL and MBedTLS implementations.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>