Build options for Linux build inside build environments
There are couple of problems when the probject is built with "NONE"
TOOLCHAIN argument for a Yocto project.

- Compile errors related with compile options: For example, undefined
symbol is occurred in OpenSSL submodule because it needs the flag
'-include base.h'.

- Cannot find -lssl and -lcrypto: In build environments, executables
need to be built and run with the sysroot of the target, not native.

- libspdm: /usr/lib/libspdm.so contains probably-redundant RPATH
/usr/lib [useless-rpaths]: CMAKE directives which add RPATH can cause
failures during do_package_qa phase in Yocto.

Signed-off-by: Gyeongtaek Lee <gt82.lee@samsung.com>