cryp: extern implmentation of libspdm_get_random_number_64
For nostd/baremetal targets, rnglib cannot currently be used. This
change allows the integrator to provide a cryptographically safe
implementation for `libspdm_get_random_number_64()`. Such that it is
abstracted away from libspdm.

This is particularly useful, when libspdm is build then statically
linked to the user application where at link time, this function can now
be resolved. Previously, you would need to build libspdm with the
implmentation present.

It is the integrators responsibility to provide an implementation that
is cryptographically safe.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>