fix default build options for generate_iv()
generate_iv() iterates over sizeof(uint64_t), accessing an array with
LIBSPDM_MAX_AEAD_IV_SIZE number of elements.  With the default build
options when no AEAD algorithms are compiled in, this array consists
of a single entry.  Gcc 12.2.0 complains about this and aborts the
compilation.  Bump the default define to 8 to allow build to succeed.

Signed-off-by: Patrick Wildt <pwildt@google.com>