CMakeLists: remove fixed options for NONE
The use of the NONE toolchain option is such that we can provide at the
build project level (buildroot etc...). However, the changes introduced
in 811f2b596def04b3a36368cf2098546d7907767f set certain compiler/linker
option that does not comply with the definition of the options as
specified in [1]. This change removes those options.

[1] https://github.com/DMTF/libspdm/blob/main/doc/build.md#linux-builds-inside-build-environments

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>