spdm_crypt_lib: Expose the base_hash_algo
The base_hash_algo is useful to know if you are using libspdm as a
library.

Reading it from the libspdm_context_t struct isn't ideal as the struct
is private and also doesn't work if using Rust for the implementation.

This commit exposes the base_hash_algo publicly using a helper
function.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>