Fix compile issue for unit test test_spdm_secured_message
In initialize_secured_message_context(), two for-loops access two arrays
with LIBSPDM_MAX_AEAD_KEY_SIZE number of elements and with
LIBSPDM_MAX_AEAD_IV_SIZE number of elements separately. Both iterations
go beyond the size of the arrays when the required macro
LIBSPDM_AEAD_AES_256_GCM_SUPPORT for the test is undefined.

Gcc 11.4.0 reports this and aborts the compilation when -DTARGET=Release
is specified with cmake for the build. Check macro define in the unit
test to avoid compile error.

Signed-off-by: Kelvin Cao <kelvin.cao@microchip.com>