Update mbedtls to 3.6.2
Mbed TLS 3.6.2 is the latest release in the long-term support 3.6.x
series. It brings in some critical security fixes and is the
recommended release of Mbed TLS from that community.

One of the fixes updates the certifi version referenced by
Mbed TLS to one that fixes this CVE:
https://nvd.nist.gov/vuln/detail/CVE-2023-37920. CVE scanning
tools pick up the version of certifi in Mbed TLS 3.6.0 and
report libspdm in violation.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>