format_nix: Add version check for uncrustify
It seems that the uncrustify config file that we have does not
work with Ubuntu 24.04 which comes with 0.78.

And the current config file with 0.78 seems to complain about
deprecated keys and actually did much more changes than what the
CI did. Hence added a version check in the format script. We need
to ensure that we change the version here as and when we start
supporting new versions of uncrustify in CI.

Signed-off-by: Manojkiran Eda <manojkiran.eda@gmail.com>