uncrustify: Upgrade the config to 0.78
This commit is an attempt to fix the uncrustify config in such a way
that it works with the latest version.

Itseems that uncrustify became a bit more smarter & a bit more stricter
w.r.t wrapping the code around 100 length. Lot of function calls in the
code base currently seems to exceed 100, hence bumping that to 120 so
that we have minimal changes during the migration.

Signed-off-by: Manojkiran Eda <manojkiran.eda@gmail.com>